Make page chrome generic so rows can reuse it
Introduces a `cortext-document` post type support flag and
`DocumentIdentity::register_for_post_type()` helper so any CPT can opt
into the shared chrome (icon meta, locked title block, identity
controls, trash mechanics). Page and dynamic row CPTs both opt in.

Renames the page-flavored pieces to their document-flavored names:
PageIdentity, PageIconBlock, PageCoverBlock, PageTrashController,
PageIdentityControls, src/blocks/page-{icon,cover}, and the
cortext_page_icon meta key all become cortext_document_*. RevisionThrottle
gates on post_type_supports('cortext-document') instead of hard-coding
the page CPT.

Extracts the shared editor body into EditorBody and the row property
panel into RowProperties so Canvas (full-page documents) and
RowDetailView (side and modal panes) render the same editor surface
for any document, with the trash banner sitting outside the locked
canvas region so the Restore button stays interactive.
