+++ includes/Plugin.php
@@ -10,21 +10,21 @@ declare( strict_types=1 );
 namespace Cortext;
 
 use Cortext\Admin\Screen;
-use Cortext\Editor\PageCoverBlock;
+use Cortext\Editor\DocumentCoverBlock;
+use Cortext\Editor\DocumentIconBlock;
 use Cortext\Editor\PageHeaderActionsBlock;
-use Cortext\Editor\PageIconBlock;
 use Cortext\Editor\RevisionThrottle;
 use Cortext\Frontend\Assets;
 use Cortext\Frontend\Template;
 use Cortext\PostType\Collection;
 use Cortext\PostType\CollectionEntries;
+use Cortext\PostType\DocumentIdentity;
 use Cortext\PostType\Field;
 use Cortext\PostType\Page;
-use Cortext\PostType\PageIdentity;
 use Cortext\PostType\PageTrashCascade;
 use Cortext\Rest\CollectionsController;
+use Cortext\Rest\DocumentTrashController;
 use Cortext\Rest\FieldsController;
-use Cortext\Rest\PageTrashController;
 use Cortext\Rest\RowsController;
 use Cortext\Rest\WorkspaceHomeController;
 use Cortext\Theming\Preferences;
@@ -43,18 +43,18 @@ final class Plugin {
 	public function boot(): void {
 		( new Screen() )->register();
 		( new Page() )->register();
-		( new PageIdentity() )->register();
+		( new DocumentIdentity() )->register();
 		( new PageTrashCascade() )->register();
 		( new Collection() )->register();
 		( new Field() )->register();
 		( new CollectionEntries() )->register();
 		( new RevisionThrottle() )->register();
-		( new PageIconBlock() )->register();
-		( new PageCoverBlock() )->register();
+		( new DocumentIconBlock() )->register();
+		( new DocumentCoverBlock() )->register();
 		( new PageHeaderActionsBlock() )->register();
 		( new CollectionsController() )->register();
 		( new FieldsController() )->register();
-		( new PageTrashController() )->register();
+		( new DocumentTrashController() )->register();
 		( new RowsController() )->register();
 		( new WorkspaceHomeController() )->register();
 		( new Template() )->register();
