+++ includes/Rest/DocumentTrashController.php
@@ -1,6 +1,11 @@
 <?php
 /**
- * REST endpoint for restoring trashed Cortext pages.
+ * REST endpoints for restoring and permanently deleting trashed Cortext
+ * documents (pages and collection rows).
+ *
+ * Page documents carry hierarchy, so restore/delete here also walks the
+ * `PageTrashCascade` marker chain to surface descendants. Row documents are
+ * flat per collection, so the cascade walk is a no-op for them.
  *
  * @package Cortext
  */
@@ -15,7 +20,7 @@ use WP_Error;
 use WP_REST_Request;
 use WP_REST_Response;
 
-final class PageTrashController {
+final class DocumentTrashController {
 
 	private const NAMESPACE = 'cortext/v1';
 
@@ -33,12 +38,12 @@ final class PageTrashController {
 
 		register_rest_route(
 			self::NAMESPACE,
-			'/pages/(?P<id>\d+)/restore',
+			'/documents/(?P<id>\d+)/restore',
 			array(
 				array(
 					'methods'             => 'POST',
 					'callback'            => array( $this, 'restore' ),
-					'permission_callback' => array( $this, 'check_trashed_page' ),
+					'permission_callback' => array( $this, 'check_document_post' ),
 					'args'                => $id_arg,
 				),
 			)
@@ -46,12 +51,12 @@ final class PageTrashController {
 
 		register_rest_route(
 			self::NAMESPACE,
-			'/pages/(?P<id>\d+)/permanent-delete',
+			'/documents/(?P<id>\d+)/permanent-delete',
 			array(
 				array(
 					'methods'             => 'POST',
 					'callback'            => array( $this, 'permanent_delete' ),
-					'permission_callback' => array( $this, 'check_trashed_page' ),
+					'permission_callback' => array( $this, 'check_document_post' ),
 					'args'                => $id_arg,
 				),
 			)
@@ -60,22 +65,23 @@ final class PageTrashController {
 
 	/**
 	 * Permission gate. Returns a `WP_Error` with status 404 when the post id
-	 * is unknown or not a `crtxt_page` so the consumer can tell "not found"
-	 * from "not allowed" without leaking permission semantics into the route
-	 * callback. WP REST honours `WP_Error` returns from permission callbacks.
+	 * is unknown or its post type does not opt into `cortext-document` so the
+	 * consumer can tell "not found" from "not allowed" without leaking
+	 * permission semantics into the route callback. WP REST honours `WP_Error`
+	 * returns from permission callbacks.
 	 *
 	 * @param WP_REST_Request $request Incoming REST request.
 	 *
 	 * @return bool|WP_Error
 	 */
-	public function check_trashed_page( WP_REST_Request $request ) {
+	public function check_document_post( WP_REST_Request $request ) {
 		$id   = (int) $request->get_param( 'id' );
 		$post = get_post( $id );
 
-		if ( ! $post || Page::POST_TYPE !== $post->post_type ) {
+		if ( ! $post || ! post_type_supports( $post->post_type, 'cortext-document' ) ) {
 			return new WP_Error(
-				'cortext_page_not_found',
-				__( 'Page not found.', 'cortext' ),
+				'cortext_document_not_found',
+				__( 'Document not found.', 'cortext' ),
 				array( 'status' => 404 )
 			);
 		}
@@ -95,8 +101,8 @@ final class PageTrashController {
 
 		if ( 'trash' !== $post->post_status ) {
 			return new WP_Error(
-				'cortext_page_not_trashed',
-				__( 'Page is not in trash.', 'cortext' ),
+				'cortext_document_not_trashed',
+				__( 'Document is not in trash.', 'cortext' ),
 				array( 'status' => 400 )
 			);
 		}
@@ -106,13 +112,16 @@ final class PageTrashController {
 		// so a post-call query alone could not tell which descendants this
 		// restore brought back versus which were already out of trash. The
 		// cascade marker is per-immediate-parent, so we walk down the chain.
-		$candidates = $this->tagged_subtree_ids( $id );
+		// Cascade is page-only; rows are flat and produce an empty subtree.
+		$candidates = Page::POST_TYPE === $post->post_type
+			? $this->tagged_subtree_ids( $id )
+			: array();
 
 		$result = wp_untrash_post( $id );
 		if ( ! $result ) {
 			return new WP_Error(
-				'cortext_page_restore_failed',
-				__( 'Page could not be restored.', 'cortext' ),
+				'cortext_document_restore_failed',
+				__( 'Document could not be restored.', 'cortext' ),
 				array( 'status' => 500 )
 			);
 		}
@@ -127,23 +136,28 @@ final class PageTrashController {
 		return new WP_REST_Response(
 			array(
 				'restored' => array_values( array_unique( array_merge( array( $id ), $revived ) ) ),
-				'post'     => $this->prepared_post( $id ),
+				'post'     => $this->prepared_post( $post ),
 			),
 			200
 		);
 	}
 
 	/**
-	 * Runs the standard `WP_REST_Posts_Controller` against the given page so
+	 * Runs the standard `WP_REST_Posts_Controller` against the given document so
 	 * the response payload matches what `useEntityRecord` already knows how to
 	 * consume. Lets clients drop a follow-up GET after a successful restore.
 	 *
-	 * @param int $id Page id to render.
+	 * @param \WP_Post $post Document post to render.
 	 *
 	 * @return mixed
 	 */
-	private function prepared_post( int $id ) {
-		$rest_request = new WP_REST_Request( 'GET', '/wp/v2/crtxt_pages/' . $id );
+	private function prepared_post( \WP_Post $post ) {
+		$post_type_object = get_post_type_object( $post->post_type );
+		$rest_base        = $post_type_object && ! empty( $post_type_object->rest_base )
+			? $post_type_object->rest_base
+			: $post->post_type;
+
+		$rest_request = new WP_REST_Request( 'GET', '/wp/v2/' . $rest_base . '/' . $post->ID );
 		$rest_request->set_param( 'context', 'edit' );
 		$response = rest_do_request( $rest_request );
 		return $response->is_error() ? null : $response->get_data();
@@ -155,16 +169,19 @@ final class PageTrashController {
 
 		if ( 'trash' !== $post->post_status ) {
 			return new WP_Error(
-				'cortext_page_not_trashed',
-				__( 'Page is not in trash.', 'cortext' ),
+				'cortext_document_not_trashed',
+				__( 'Document is not in trash.', 'cortext' ),
 				array( 'status' => 400 )
 			);
 		}
 
 		// Snapshot the subtree up front, then delete leaves-first so WP's
 		// hierarchical-delete reparenting (`post_parent = $deleted->post_parent`)
-		// has nothing to do for any non-leaf.
-		$descendants = $this->tagged_subtree_ids( $id );
+		// has nothing to do for any non-leaf. Cascade is page-only; rows are
+		// flat and produce an empty descendant list.
+		$descendants = Page::POST_TYPE === $post->post_type
+			? $this->tagged_subtree_ids( $id )
+			: array();
 
 		$deleted = array();
 		foreach ( array_reverse( $descendants ) as $descendant_id ) {
@@ -175,8 +192,8 @@ final class PageTrashController {
 
 		if ( ! wp_delete_post( $id, true ) ) {
 			return new WP_Error(
-				'cortext_page_delete_failed',
-				__( 'Page could not be deleted.', 'cortext' ),
+				'cortext_document_delete_failed',
+				__( 'Document could not be deleted.', 'cortext' ),
 				array( 'status' => 500 )
 			);
 		}
@@ -194,7 +211,7 @@ final class PageTrashController {
 	 * BFS over the cascade-marker tree rooted at `$root_id`, returning every
 	 * trashed descendant whose marker chain ties back to the root. The marker
 	 * stores each child's immediate parent, so the walk expands one level at
-	 * a time.
+	 * a time. Page-only by design.
 	 *
 	 * @param int $root_id Page id to walk from.
 	 *
