+++ src/blocks/index.js
@@ -1,6 +1,6 @@
 // Central entry for Cortext-provided blocks. Keeps registration side-effects
 // in one place so src/index.js stays a thin bootstrap.
 import './data-view';
-import './page-icon';
-import './page-cover';
+import './document-icon';
+import './document-cover';
 import './page-header-actions';
