+++ tests/php/test-document-identity.php
@@ -1,6 +1,6 @@
 <?php
 /**
- * Tests for Cortext\PostType\PageIdentity.
+ * Tests for Cortext\PostType\DocumentIdentity.
  *
  * @package Cortext
  */
@@ -9,21 +9,28 @@ declare( strict_types=1 );
 
 namespace Cortext\Tests;
 
+use Cortext\PostType\DocumentIdentity;
 use Cortext\PostType\Page;
-use Cortext\PostType\PageIdentity;
 use WorDBless\BaseTestCase;
 
-final class Test_Page_Identity extends BaseTestCase {
+final class Test_Document_Identity extends BaseTestCase {
+
+	protected function setUp(): void {
+		parent::setUp();
+		// The prepender now gates on post_type_supports('cortext-document'),
+		// so the test post type needs the trait wired up before each case.
+		DocumentIdentity::register_for_post_type( Page::POST_TYPE );
+	}
 
 	public function test_header_blocks_markup_contains_title_only(): void {
-		$markup = PageIdentity::header_blocks_markup();
+		$markup = DocumentIdentity::header_blocks_markup();
 
 		$this->assertStringContainsString( '<!-- wp:post-title', $markup );
 		$this->assertStringNotContainsString( 'cortext/page-header-actions', $markup );
 	}
 
 	public function test_prepend_header_blocks_strips_legacy_actions_on_update(): void {
-		$identity = new PageIdentity();
+		$identity = new DocumentIdentity();
 		$content  = '<!-- wp:cortext/page-header-actions {"lock":{"move":true,"remove":true}} /--><!-- wp:paragraph --><p>Body</p><!-- /wp:paragraph -->';
 
 		$data = $identity->prepend_header_blocks(
@@ -40,7 +47,7 @@ final class Test_Page_Identity extends BaseTestCase {
 	}
 
 	public function test_prepend_header_blocks_adds_title_without_legacy_actions_on_create(): void {
-		$identity = new PageIdentity();
+		$identity = new DocumentIdentity();
 		$content  = '<!-- wp:cortext/page-header-actions {"lock":{"move":true,"remove":true}} /--><!-- wp:paragraph --><p>Body</p><!-- /wp:paragraph -->';
 
 		$data = $identity->prepend_header_blocks(
@@ -58,8 +65,8 @@ final class Test_Page_Identity extends BaseTestCase {
 	}
 
 	public function test_prepend_header_blocks_does_not_duplicate_existing_serialized_title(): void {
-		$identity = new PageIdentity();
-		$content  = PageIdentity::header_blocks_markup() . '<!-- wp:paragraph --><p>Body</p><!-- /wp:paragraph -->';
+		$identity = new DocumentIdentity();
+		$content  = DocumentIdentity::header_blocks_markup() . '<!-- wp:paragraph --><p>Body</p><!-- /wp:paragraph -->';
 
 		$data = $identity->prepend_header_blocks(
 			array(
@@ -71,5 +78,21 @@ final class Test_Page_Identity extends BaseTestCase {
 
 		$unslashed = wp_unslash( $data['post_content'] );
 		$this->assertSame( 1, substr_count( $unslashed, '<!-- wp:post-title' ) );
+	}
+
+	public function test_prepend_header_blocks_skips_post_types_without_document_support(): void {
+		$identity = new DocumentIdentity();
+		$content  = '<!-- wp:paragraph --><p>Body</p><!-- /wp:paragraph -->';
+
+		$data = $identity->prepend_header_blocks(
+			array(
+				'post_type'    => 'post',
+				'post_content' => wp_slash( $content ),
+			),
+			array()
+		);
+
+		$unslashed = wp_unslash( $data['post_content'] );
+		$this->assertStringNotContainsString( '<!-- wp:post-title', $unslashed );
 	}
 }
